auth/requirements.txt
aiohttp==3.7.3
aiohttp-oauth2==0.0.5
aioredis==1.3.1
alembic==1.5.3
SQLAlchemy==1.4.0b2
asyncpg==0.21.0
PyJWT==2.0.1
cryptography==3.3.1
pycryptodomex==3.9.9
psycopg2-binary==2.8.6
py-auth-header-parser==1.0.2